requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
django-cors-headers
gunicorn
celery==5.3.6
redis==5.0.1
django-celery-beat==2.6.0
httpx==0.27.0
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.2
drf-yasg
